+++ tests/Transformable.php
@@ -1,0 +1,39 @@
+<?php
+
+/**
+ * Recipe.
+ *
+ * LICENSE
+ *
+ * This source file is subject to the MIT license
+ * license that are bundled with this package in the folder licences
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to richarddeloge@gmail.com so we can send you a copy immediately.
+ *
+ *
+ * @copyright   Copyright (c) 2009-2021 EIRL Richard Déloge (richarddeloge@gmail.com)
+ * @copyright   Copyright (c) 2020-2021 SASU Teknoo Software (https://teknoo.software)
+ *
+ * @link        http://teknoo.software/recipe Project website
+ *
+ * @license     http://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richarddeloge@gmail.com>
+ */
+
+namespace Teknoo\Tests\Recipe;
+
+use Teknoo\Recipe\Ingredient\TransformableInterface;
+
+class Transformable implements TransformableInterface
+{
+    public function __construct(
+        private array $values,
+    ) {
+    }
+
+    public function transform(): mixed
+    {
+        return $this->values;
+    }
+}